Practical 1: add relating Book and Author
This additional exercise can be tricky, as we need to manage
bidirectional cross-references between authors and books without
having a database do it for us. This commit adds an example
solution for this, which makes the book side the "owning" side
(i.e. the side you modify) - this is similar to what we end
up doing with a database on Practical 2.